// src/types/auth.ts
export type TUser = {
    firstName: string;
    middleName?: string;
    lastName: string;
    email: string;
    userName: string;
    password: string;
};

export type TUserState = {
    firstName: string;
    middleName?: string;
    lastName: string;
    email: string;
    userName: string;
    password: string;
    verifyPassword: string;
};